+++ DownloaderForReddit/Extractors/ImgurExtractor.py
@@ -193,7 +193,7 @@ class ImgurExtractor(BaseExtractor):
                     url = picture.mp4
                     extension = 'mp4'
             self.make_content(url, file_name, extension)
-        except (AttributeError, NameError):
+        except (AttributeError, NameError, TypeError):
             message = 'Unrecognized extension'
             self.handle_failed_extract(message=message, extractor_error_message=message)