+++ DownloaderForReddit/GUI/DownloaderForRedditGUI.py
@@ -46,6 +46,7 @@ from ..Utils.Exporters import TextExporter, JsonExporter, XMLExporter
 from ..ViewModels.RedditObjectListModel import RedditObjectListModel
 from ..GUI.AddRedditObjectDialog import AddUserDialog
 from ..GUI.FfmpegInfoDialog import FfmpegInfoDialog
+from ..GUI.ExistingRedditObjectAddDialog import ExistingRedditObjectAddDialog
 from ..version import __version__
 
 
@@ -219,6 +220,8 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
             user_menu.addSeparator()
             download_single = user_menu.addAction('Download %s' % user.name)
             download_single.triggered.connect(lambda: self.run_single_user((user, None)))
+            if self.running:
+                download_single.setEnabled(False)
 
         add_user.triggered.connect(self.add_user_dialog)
         remove_user.triggered.connect(self.remove_user)
@@ -264,6 +267,8 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
             subreddit_menu.addSeparator()
             download_single = subreddit_menu.addAction('Download %s' % subreddit.name)
             download_single.triggered.connect(lambda: self.run_single_subreddit((subreddit, None)))
+            if self.running:
+                download_single.setEnabled(False)
 
         add_subreddit.triggered.connect(self.add_subreddit_dialog)
         remove_subreddit.triggered.connect(self.remove_subreddit)
@@ -548,7 +553,7 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
         elif text.startswith('Saved'):
             self.update_status_bar_download_count()
             self.output_box.append(text)
-        elif text.startswith('Count'):
+        elif text.startswith('$$Count'):
             t, count = text.rsplit(' ', 1)
             self.download_count += int(count)
         else:
@@ -714,7 +719,7 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
             user = self.make_user(new_user)
             reply = self.add_reddit_object_to_list(user, self.user_list_model)
             if reply == 'NAME_EXISTS':
-                Message.name_in_list(self, new_user)
+                self.handle_existing_user_name(new_user)
             elif reply == 'INVALID_NAME':
                 Message.not_valid_name(self, new_user)
             else:
@@ -722,6 +727,36 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
                 self.check_user_download_on_add(user)
         except KeyError:
             Message.no_user_list(self)
+
+    def handle_existing_user_name(self, user_name):
+        """
+        Handles when an existing user name is added to the list.  If the settings manager indicates that users should
+        be downloaded on add, a dialog is used to ascertain if the app user wants to run a single download for the
+        user they are attempting to add.
+        :param user_name: The name of the existing user.
+        """
+        if not self.running and self.settings_manager.download_users_on_add:
+            existing_user_dialog = ExistingRedditObjectAddDialog(user_name, 'USER')
+            dialog = existing_user_dialog.exec_()
+            if dialog == QtWidgets.QDialog.Accepted:
+                self.download_existing_user(user_name)
+        else:
+            Message.name_in_list(self, user_name)
+
+    def download_existing_user(self, user_name):
+        """
+        Gets an existing User object from the current list that has the supplied user_name, then runs a single download
+        for that user.
+        :param user_name: The name of the user for which a single download is to be run
+        """
+        current_list = self.get_working_list('USER')
+        user = None
+        for item in current_list.reddit_object_list:
+            if item.name == user_name:
+                user = item
+                break
+        if user is not None:
+            self.run_single_user((user, None))
 
     def check_user_download_on_add(self, user):
         """
@@ -891,6 +926,11 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
                                                                  'removal_reason': reason})
 
     def get_working_list(self, object_type):
+        """
+        Returns the list that is currently being displayed based on the supplied object type.
+        :param object_type: The type of list that is to be returned.
+        :return: The List model that of the supplied object type that is currently being displayed.
+        """
         if object_type == 'USER':
             return self.user_list_model
         else:
@@ -925,6 +965,36 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
                 self.check_subreddit_download_on_add(subreddit)
         except KeyError:
             Message.no_user_list(self)
+
+    def handle_existing_subreddit_name(self, sub_name):
+        """
+        Handles when an existing subreddit name is added to the list.  If the settings manager indicates that subreddits
+        should be downloaded on add, a dialog is used to ascertain if the app user wants to run a single download for
+        the subreddit they are attempting to add.
+        :param sub_name: The name of the existing subreddit.
+        """
+        if not self.running and self.settings_manager.download_subreddits_on_add:
+            existing_sub_dialog = ExistingRedditObjectAddDialog(sub_name, 'SUBREDDIT')
+            dialog = existing_sub_dialog.exec_()
+            if dialog == QtWidgets.QDialog.Accepted:
+                self.download_existing_subreddit(sub_name)
+        else:
+            Message.name_in_list(self, sub_name)
+
+    def download_existing_subreddit(self, sub_name):
+        """
+        Gets an existing Subreddit object from the current list that has the supplied sub_name, then runs a single
+        download for that subreddit.
+        :param sub_name: The name of the subreddit for which a single download is to be run.
+        """
+        current_list = self.get_working_list('SUBREDDIT')
+        sub = None
+        for item in current_list.reddit_object_list:
+            if item.name == sub_name:
+                sub = item
+                break
+        if sub is not None:
+            self.run_single_subreddit((sub, None))
 
     def check_subreddit_download_on_add(self, subreddit):
         """
@@ -1109,6 +1179,7 @@ class DownloaderForRedditGUI(QtWidgets.QMainWindow, Ui_MainWindow):
         self.download_count = 0
         self.output_box.clear()
         self.download_button.setText('Downloading...Click to Stop Download')
+        self.statusbar.clearMessage()
         self.add_user_button.setDisabled(True)
         self.remove_user_button.setDisabled(True)
         self.add_subreddit_button.setDisabled(True)