+++ DownloaderForReddit/GUI/FailedDownloadsDialog.py
@@ -22,6 +22,7 @@ You should have received a copy of the GNU General Public License
 along with Downloader for Reddit.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+from os import path
 
 from PyQt5 import QtCore, QtWidgets
 from PyQt5.QtGui import QCursor
@@ -72,33 +73,38 @@ class FailedDownloadsDialog(QtWidgets.QDialog, Ui_failed_downloads_dialog):
 
     def display_context_menu(self):
         menu = QtWidgets.QMenu()
-        index = self.table_view.selectedIndexes()[0].row()
-        post = self.table_model.data_list[index]
-        visit_author_reddit = menu.addAction('Visit Author Page')
-        visit_subreddit = menu.addAction('Visit Subreddit')
-        visit_post_page = menu.addAction('Visit Post')
-        menu.addSeparator()
+        selected_indexes = self.table_view.selectedIndexes()
+        if len(selected_indexes) > 0:            
+            index = selected_indexes[0].row()
+            post = self.table_model.data_list[index]
+            
+            visit_author_reddit = menu.addAction('Visit Author Page')
+            visit_subreddit = menu.addAction('Visit Subreddit')
+            visit_post_page = menu.addAction('Visit Post')
+
+            visit_author_reddit.triggered.connect(lambda: SystemUtil.open_in_system('www.reddit.com/u/%s' % post.author))
+            visit_subreddit.triggered.connect(lambda: SystemUtil.open_in_system('www.reddit.com/r/%s' % post.subreddit))
+            visit_post_page.triggered.connect(lambda: SystemUtil.open_in_system(post.url))
+            menu.addSeparator()
+
         export_url_list = menu.addAction('Export Url List')
+        export_url_list.triggered.connect(self.export_url_list)
         menu.addSeparator()
-        export_selected_menu = menu.addMenu('Export Selected')
-        export_selected_as_text = export_selected_menu.addAction('Export As Text')
-        export_selected_as_json = export_selected_menu.addAction('Export As Json')
-        export_selected_as_xml = export_selected_menu.addAction('Export As Xml')
+
+        if len(selected_indexes) > 0: 
+            export_selected_menu = menu.addMenu('Export Selected')
+            export_selected_as_text = export_selected_menu.addAction('Export As Text')
+            export_selected_as_json = export_selected_menu.addAction('Export As Json')
+            export_selected_as_xml = export_selected_menu.addAction('Export As Xml')
+
+            export_selected_as_text.triggered.connect(self.export_selected_to_text)
+            export_selected_as_json.triggered.connect(self.export_selected_to_json)
+            export_selected_as_xml.triggered.connect(self.export_selected_to_xml)
 
         export_all_menu = menu.addMenu('Export All')
         export_list_as_text = export_all_menu.addAction('Export As Text')
         export_list_as_json = export_all_menu.addAction('Export As Json')
-        export_list_as_xml = export_all_menu.addAction('Export As Xml')
-
-        visit_author_reddit.triggered.connect(lambda: SystemUtil.open_in_system('www.reddit.com/u/%s' % post.author))
-        visit_subreddit.triggered.connect(lambda: SystemUtil.open_in_system('www.reddit.com/r/%s' % post.subreddit))
-        visit_post_page.triggered.connect(lambda: SystemUtil.open_in_system(post.url))
-
-        export_url_list.triggered.connect(self.export_url_list)
-
-        export_selected_as_text.triggered.connect(self.export_selected_to_text)
-        export_selected_as_json.triggered.connect(self.export_selected_to_json)
-        export_selected_as_xml.triggered.connect(self.export_selected_to_xml)
+        export_list_as_xml = export_all_menu.addAction('Export As Xml')    
 
         export_list_as_text.triggered.connect(self.export_list_to_text)
         export_list_as_json.triggered.connect(self.export_list_to_json)
@@ -135,43 +141,48 @@ class FailedDownloadsDialog(QtWidgets.QDialog, Ui_failed_downloads_dialog):
             else:
                 self.close()
 
+    def get_selected_rows(self):
+        return [self.table_model.data_list[x.row()] for x in self.table_view.selectionModel().selectedRows()]
+
     def export_url_list(self):
         url_list = [post.url for post in self.table_model.data_list]
         file_path = self.get_file_path('failed_urls.txt', 'Text Files (*.txt)')
-        TextExporter.export_url_list(url_list, file_path)
+        if file_path:
+            TextExporter.export_url_list(url_list, file_path)
 
     def export_selected_to_text(self):
         file_path = self.get_file_path('failed_urls.txt', 'Text Files (*.txt)')
-        TextExporter.export_url_list([self.table_model.data_list[x.row()] for x in
-                                      self.table_view.selectionModel().selectedRows()],
-                                     file_path)
+        if file_path:
+            TextExporter.export_url_list(self.get_selected_rows(),file_path)
 
     def export_list_to_text(self):
         file_path = self.get_file_path('failed_downloads.txt', 'Text Files (*.txt)')
-        TextExporter.export_posts_to_text(self.table_model.data_list, file_path)
+        if file_path:
+            TextExporter.export_posts_to_text(self.table_model.data_list, file_path)
 
     def export_selected_to_json(self):
         file_path = self.get_file_path('failed_downloads.json', 'Json Files (*.json)')
-        JsonExporter.export_posts_to_json([self.table_model.data_list[x.row()] for x in
-                                           self.table_view.selectionModel().selectedRows()],
-                                          file_path)
+        if file_path:
+            JsonExporter.export_posts_to_json(self.get_selected_rows(),file_path)
 
     def export_list_to_json(self):
         file_path = self.get_file_path('failed_downloads.json', 'Json Files (*.json)')
-        JsonExporter.export_posts_to_json(self.table_model.data_list, file_path)
+        if file_path:
+            JsonExporter.export_posts_to_json(self.table_model.data_list, file_path)
 
     def export_selected_to_xml(self):
         file_path = self.get_file_path('failed_downloads.xml', 'Xml Files (*.xml)')
-        XMLExporter.export_posts_to_xml([self.table_model.data_list[x.row()] for x in
-                                         self.table_view.selectionModel().selectedRows()],
-                                        file_path)
+        if file_path:
+            XMLExporter.export_posts_to_xml(self.get_selected_rows(),file_path)
 
     def export_list_to_xml(self):
         file_path = self.get_file_path('failed_downloads.xml', 'Xml Files (*.xml)')
-        XMLExporter.export_posts_to_xml(self.table_model.data_list, file_path)
+        if file_path:
+            XMLExporter.export_posts_to_xml(self.table_model.data_list, file_path)
 
     def get_file_path(self, suggested_name, extension):
+        suggested_path = path.join(self.settings_manager.save_directory, suggested_name)
         file_path, _ = QtWidgets.QFileDialog.getSaveFileName(self, 'Select Save Path',
-                                                             self.settings_manager.save_directory + suggested_name,
+                                                             suggested_path,
                                                              extension)
         return file_path