+++ DownloaderForReddit/Utils/Injector.py
@@ -24,7 +24,6 @@ along with Downloader for Reddit.  If not, see <http://www.gnu.org/licenses/>.
 from queue import Queue
 
 from ..Persistence.SettingsManager import SettingsManager
-from ..Database.DatabaseHandler import DatabaseHandler
 
 
 settings_manager = None
@@ -42,6 +41,7 @@ def get_settings_manager():
 def get_database_handler():
     global database_handler
     if database_handler is None:
+        from ..Database.DatabaseHandler import DatabaseHandler
         database_handler = DatabaseHandler()
     return database_handler